refactor: move commands code out of main.rs
