+++ src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod check;
+pub mod download;
+pub mod install;
+pub mod list;
+pub mod make_default;
