+++ src/main.rs
@@ -1,20 +1,19 @@
 use std::io::Write;
-use std::path::PathBuf;
-use std::{fs::File, path::Path};
 
 use clap::{Parser, Subcommand};
 use clap_verbosity_flag::{InfoLevel, Verbosity};
 use lazy_static::lazy_static;
-use log::{debug, error, info, warn};
+use log::{debug, error, info};
 use regex::Regex;
 
 mod archives;
+mod commands;
 mod filesys;
 mod github;
 mod models;
 use github::client::{get_asset, get_release};
 use poof::is_env_compatible;
-use semver_utils::{SemverStringConversion, SemverStringPrefix};
+use semver_utils::SemverStringConversion;
 mod platform_info;
 mod semver_utils;
 mod utils;
@@ -140,7 +139,11 @@ fn main() {
 
             let release = get_release(&args.repo, args.tag.as_deref());
             let binary = get_asset(&release, is_env_compatible);
-            download_binary(binary.name(), binary.browser_download_url(), &current_dir);
+            commands::download::download_binary(
+                binary.name(),
+                binary.browser_download_url(),
+                &current_dir,
+            );
         }
         Cmd::Install(args) => {
             info!(
@@ -148,7 +151,7 @@ fn main() {
                 &args.repo,
                 args.tag.as_deref().unwrap_or("(latest)")
             );
-            process_install(&args.repo, args.tag.as_deref());
+            commands::install::process_install(&args.repo, args.tag.as_deref());
         }
         Cmd::Use(args) => {
             let version = args.tag.as_deref().unwrap_or("latest");
@@ -156,11 +159,14 @@ fn main() {
                 "Setting version '{}' as default for {}",
                 version, &args.repo
             );
-            make_default(&args.repo, version);
+            if let Err(e) = commands::make_default::set_default(&args.repo, version) {
+                error!("Failed to set default version: {}", e);
+                std::process::exit(110);
+            }
             info!("Version '{}' set as default.", version);
         }
         Cmd::List => {
-            let list = filesys::list_installed_assets();
+            let list = commands::list::list_installed_assets();
             if list.is_empty() {
                 info!("No installed binaries found.");
             } else {
@@ -181,269 +187,13 @@ fn main() {
             }
         }
         Cmd::Check => {
-            check_if_bin_in_path();
+            commands::check::check_if_bin_in_path();
         }
         Cmd::Version => {
             println!("{}", platform_info::long_version());
         }
         Cmd::Debug => {
             platform_info::debug_info();
-        }
-    }
-}
-
-fn process_install(repo: &str, tag: Option<&str>) {
-    // let config_dir = filesys::get_config_dir().ok_or(libc::ENOENT).unwrap();
-    // info!("Config directory: {}", config_dir);
-    let cache_dir: PathBuf = filesys::get_cache_dir().ok_or(libc::ENOENT).unwrap();
-    debug!("Cache directory: {}", cache_dir.display());
-
-    // download binary
-    let release = get_release(repo, tag);
-    let binary = get_asset(&release, is_env_compatible);
-    let download_to = get_install_path(&cache_dir, repo, release.tag_name());
-    download_binary(binary.name(), binary.browser_download_url(), &download_to);
-
-    // extract binary
-    let archive_path = download_to.join(binary.name());
-    archives::extract_to_dir_depending_on_content_type(
-        binary.content_type(),
-        &archive_path,
-        &download_to,
-    )
-    .expect("Failed to extract archive");
-    debug!("Extracted to: {}", download_to.display());
-
-    // install binary
-    let version: String = release.tag_name().strip_v();
-    install_binary(&archive_path, repo, &version);
-    info!("{} installed successfully.", binary.name());
-    check_if_bin_in_path();
-    std::process::exit(0);
-}
-
-fn install_binary(archive_path: &Path, repo: &str, version: &str) {
-    let data_dir: PathBuf = filesys::get_data_dir().ok_or(libc::ENOENT).unwrap();
-    debug!("Data directory: {}", data_dir.display());
-    let install_dir: PathBuf = get_install_path(&data_dir, repo, version);
-    prepare_install_dir(&install_dir);
-
-    let execs_to_install: Vec<PathBuf> =
-        filesys::find_exec_files_from_extracted_archive(archive_path);
-    for exec in execs_to_install {
-        install_executable(&exec, &install_dir);
-    }
-}
-
-fn prepare_install_dir(install_dir: &PathBuf) {
-    debug!("Installing to: {}", install_dir.display());
-    // Create the installation directory if it doesn't exist
-    if !install_dir.exists() {
-        std::fs::create_dir_all(install_dir).unwrap();
-    } else if install_dir.is_dir() && install_dir.read_dir().unwrap().count() > 0 {
-        // Check if the directory is not empty
-        // If it is not empty, warn the user and exit
-        warn!(
-            "Version is already installed. Check content in {} dir.",
-            install_dir.display()
-        );
-        warn!("If you want to reinstall, please remove the directory first.");
-        std::process::exit(0);
-    } // else overwrite empty dir of possibly left-over dumb file with dir name
-}
-
-fn install_executable(exec: &PathBuf, install_dir: &Path) {
-    let file_name = exec.file_name().unwrap();
-    let installed_exec = install_dir.join(file_name);
-    debug!("Copying {} to {}", exec.display(), installed_exec.display());
-    if let Err(e) = std::fs::copy(exec, &installed_exec) {
-        error!(
-            "Error copying {} to {}: {}",
-            exec.display(),
-            installed_exec.display(),
-            e
-        );
-        error!("Installation failed.");
-        std::process::exit(103);
-    }
-    debug!("Making {} executable", file_name.to_string_lossy());
-    // Set executable permissions, platform-specific
-    // Note: Windows does not require setting executable permissions
-    #[cfg(not(target_os = "windows"))]
-    {
-        // Make the file executable on Unix-like systems
-        filesys::make_executable(&installed_exec);
-        // Create a symlink in the bin directory
-        let bin_dir: PathBuf = filesys::get_bin_dir().ok_or(libc::ENOENT).unwrap();
-        let symlink_path = bin_dir.join(file_name);
-        debug!(
-            "Creating symlink {} -> {}",
-            symlink_path.display(),
-            installed_exec.display()
-        );
-        // check if symlink already exists
-        if symlink_path.exists() {
-            warn!(
-                "Symlink {} already exists. Skipping.",
-                symlink_path.display()
-            );
-            return;
-        }
-        // if symlink does not exist, create it to make exec available in PATH
-        if let Err(e) = filesys::symlink(&installed_exec, &symlink_path) {
-            error!(
-                "Cannot symlink {} -> {}\n\nInstallation failed. {}",
-                symlink_path.display(),
-                installed_exec.display(),
-                e
-            );
-        } else {
-            info!(
-                "Symlink created: {} -> {}",
-                symlink_path.display(),
-                installed_exec.display()
-            );
-        }
-    }
-}
-
-// Function to handle downloading and potentially installing binaries
-fn download_binary(filename: &String, download_url: &String, download_to: &PathBuf) {
-    info!("Downloading {}\nfrom {}", filename, download_url);
-    let response = reqwest::blocking::get(download_url).unwrap();
-    if response.status().is_success() {
-        // Ensure the directory exists
-        std::fs::create_dir_all(download_to).unwrap();
-
-        // Create the file path and open it for writing
-        let archive_path = download_to.join(filename);
-        let mut file = File::create(&archive_path).unwrap();
-
-        debug!("Saving to: {}", archive_path.display());
-        std::io::copy(&mut response.bytes().unwrap().as_ref(), &mut file).unwrap();
-        info!("Download complete.");
-    } else {
-        error!("Download failed!");
-        std::process::exit(99)
-    }
-}
-
-fn get_install_path(base: &Path, repo: &str, version: &str) -> PathBuf {
-    // Convert repo path to filesystem-friendly format by replacing '/' with OS separator
-    let repo_path = repo.replace('/', std::path::MAIN_SEPARATOR_STR);
-    // Creating path as: base_dir/username/reponame/version
-    base.join(&repo_path).join(version)
-}
-
-fn check_if_bin_in_path() {
-    let bin_dir: PathBuf = filesys::get_bin_dir().ok_or(libc::ENOENT).unwrap();
-    let position = platform_info::check_dir_in_path(bin_dir.to_str().unwrap());
-    match position {
-        -1 => {
-            warn!("Bin directory not found in PATH.");
-            warn!(
-                "Please add {} to your PATH. For example, run: \n\n{}\n",
-                bin_dir.display(),
-                get_export_command()
-            );
-            warn!("This is required to run the installed binaries.");
-        }
-        0 => debug!("Everything looks good. Bin directory is the first in PATH."),
-        _ => {
-            warn!("Bin directory is not the first in PATH.");
-            warn!(
-                "Please move {} to the beginning of your PATH.",
-                bin_dir.display()
-            );
-        }
-    }
-}
-
-fn get_export_command() -> String {
-    let bin_dir: PathBuf = filesys::get_bin_dir().ok_or(libc::ENOENT).unwrap();
-    format!("export PATH=\"{}:$PATH\"", bin_dir.to_str().unwrap())
-}
-
-fn make_default(repo: &str, version: &str) {
-    let data_dir: PathBuf = filesys::get_data_dir().ok_or(libc::ENOENT).unwrap();
-    let wanted_dir: PathBuf = get_install_path(&data_dir, repo, version);
-    if !wanted_dir.exists() {
-        error!(
-            "Version {} of repository '{}' not installed. Quitting.",
-            version, repo
-        );
-        std::process::exit(110);
-    }
-    // Get the bin directory
-    let bin_dir: PathBuf = filesys::get_bin_dir().ok_or(libc::ENOENT).unwrap();
-    // List all executable files in the directory with the wanted version
-    let entries = match std::fs::read_dir(&wanted_dir) {
-        Ok(entries) => entries,
-        Err(e) => {
-            error!("Failed to read directory {}: {}", wanted_dir.display(), e);
-            std::process::exit(1);
-        }
-    };
-    // Process each binary in wanted_dir
-    for entry in entries {
-        let entry = match entry {
-            Ok(entry) => entry,
-            Err(e) => {
-                warn!("Failed to read entry: {}", e);
-                continue;
-            }
-        };
-
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-
-        // Check if the file is executable
-        #[cfg(not(target_os = "windows"))]
-        if !filesys::is_executable(&path) {
-            continue;
-        }
-
-        let file_name = match path.file_name() {
-            Some(name) => name,
-            None => continue,
-        };
-
-        let symlink_path = bin_dir.join(file_name);
-
-        // Delete existing symlink if it exists
-        if symlink_path.exists() {
-            info!("Removing existing symlink: {}", symlink_path.display());
-            if let Err(e) = std::fs::remove_file(&symlink_path) {
-                warn!(
-                    "Failed to remove existing symlink {}: {}",
-                    symlink_path.display(),
-                    e
-                );
-                continue;
-            }
-        }
-
-        // Create new symlink
-        debug!(
-            "Creating symlink {} -> {}",
-            symlink_path.display(),
-            path.display()
-        );
-        if let Err(e) = filesys::symlink(&path, &symlink_path) {
-            error!(
-                "Failed to create symlink {} -> {}: {}",
-                symlink_path.display(),
-                path.display(),
-                e
-            );
-        } else {
-            info!(
-                "Symlink created: {} -> {}",
-                symlink_path.display(),
-                path.display()
-            );
         }
     }
 }
